Stream now loads behind a (placeholder) splash screen
More details: The MediaPlayer now runs on a Service, and prepares behind
a splash screen so that the stream can be immediately started on the
main screen. This is still quite rough; there are some lifecycle issues
with the Service/MediaPlayer (the service appears to restart upon exit,
for example). Requires more testing and polish, but the general func-
tionality is there.

Minor changes:
  - Renamed package, need to rename others to follow convention

Signed-off-by: Arjun Raghavan <arjun.raghavan.2k@gmail.com>