Added "pre-splash" screen for seamless start-up
More details: The main splash screen has to inflate a layout, start a
service, and listen for a Broadcast, so there was a little lag on
startup--users would see a blank white screen BEFORE the splash screen,
which is unacceptable. Now, there's a pre-splash screen with only the
logo displayed that starts up almost immediately and then hands off to
the main splash screen when the latter's ready. (For now, this pre-
splash screen should only work for devices on Android 6.0+ because of
compatibility issues, but I still need to test that older versions jump
straight to the main splash screen as required).

NOTE: Errors are being reported in app/res/values/styles.xml but they
don't seem to be impacting the app or causing build errors in any way,
so I'm just leaving them alone.

Signed-off-by: Arjun Raghavan <arjun.raghavan.2k@gmail.com>