Added some basic stream error handling
More details: Now, on startup, if the app is connected to a WiFI network
that blocks the Radio stream (right now this list only includes UCLA_
WEB and UCLA_WEB_RES), it will prompt the user to switch networks. Also,
if the stream drops connection temporarily, the app will now attempt to
reconnect and continue (before, it would simply treat the end of the
stream as the end of the MediaPlayer's input and loop whatever was
recorded).

Signed-off-by: Arjun Raghavan <arjun.raghavan.2k@gmail.com>