Added new launcher icons; changed notification icon; fixed some bugs.
More details: New notification icon is just the favicon for the website.
Added a color (trianglified) version of the old launcher icon as well as
a circle bear icon (for fun I guess?).

Bug fixes: If the DJ or Schedule list fail to update for some reason,
they will now retry instead of just leaving a blank Fragment. Also,
SplashActivity was stopping StreamService whenever it was destroyed;
this was a problem if the user stopped the activity while it was
loading, because then the MediaPlayer was killed and not revived,
leading to a NullPointerException when isPlaying() was called. This has
been fixed.

Signed-off-by: Arjun Raghavan <arjun.raghavan.2k@gmail.com>