Changed Shows fragment UI to a 2-column grid
More details: Just think this looks a little nicer; I've left the old
layout file/code just in case people prefer the old style. Also: missing
show art now defaults to the same art as on the website, which looks
better anyway, and logo_banner_white.png was shrunk by 50% because it
was so large that there was an occasional OOM error thrown by the app on
start-up.

Signed-off-by: Arjun Raghavan <arjun.raghavan.2k@gmail.com>