Made play/pause button a little nicer-looking (it's a circle now)
More details: Truthfully, the circle button as it is right now looks
a little out of place. That's mainly because it's in preparation for
a revamp of the streaming UI, and it doesn't look TOO bad anyway. Also,
it changes color with the show art as well.

Minor additions: Actual play/pause icons were added (obtained from
material.io/tools/icons) because the default ones are ugly as hell.

Minor fixes: OOM errors were still cropping up occasionally, so now
if the app can't load the banner logo on the UI thread it'll delegate
it to Picasso to load it aynchronously. This will create some visual
lag, but it happens rarely enough that it shouldn't be too bad an issue.

Signed-off-by: Arjun Raghavan <arjun.raghavan.2k@gmail.com>