Fixed handling of connection drops; stream resumes automatically.
More details: In addition, there's a visual indication of a stream drop
now so that the user doesn't think something's wrong with the app or
the server. Also, if the user pauses the stream and then hits play, it
resumes from the point where they hit pause, and NOT the actual current
position of the stream instead. After quite a bit of grueling work, upon
testing against the actual website it turns out this is because of the
way SHOUTcast works, and has nothing to do with the app. Unfortunately,
this means that if the user has the app open and paused for a while it
will run quite out of sync with the server unless they restart it.

Signed-off-by: Arjun Raghavan <arjun.raghavan.2k@gmail.com>